abstract out displayFAQ
